+++ src/lib/components/lemmy/moderation/CommentModerationMenu.svelte
@@ -2,12 +2,8 @@
     
     import type {
         CommentView,
-        Community,
-        CommunityView,
-        PostView,
     } from 'lemmy-js-client'
-    import { amMod, ban, isAdmin, remove } from './moderation'
-    import { Color } from '$lib/ui/colors.js'
+    import { amMod, isAdmin } from './moderation'
     import { getClient } from '$lib/lemmy'
     import { isCommentView } from '$lib/lemmy/item.js'
     import { profile } from '$lib/auth.js'
@@ -15,8 +11,11 @@
     import { userSettings } from '$lib/settings'
 
     import Menu from '$lib/components/ui/menu/Menu.svelte'
+    import BanCommunityModal from '$lib/components/lemmy/moderation/BanCommunityModal.svelte'
+    import BanInstanceModal from '$lib/components/lemmy/moderation/BanInstanceModal.svelte'
     import Button from '$lib/components/input/Button.svelte'
     import MenuButton from '$lib/components/ui/menu/MenuButton.svelte'
+    import RemoveModal from './RemoveModal.svelte'
     import ShieldIcon from '$lib/components/lemmy/moderation/ShieldIcon.svelte'
 
     import { 
@@ -30,6 +29,10 @@
 
     export let item: CommentView
 
+    let removing = false
+    let banningInstance = false
+    let banningCommunity = false
+    let purging = false
 
     const distinguish = async function(comment:CommentView) {
         let distinguished: boolean = comment.comment.distinguished;
@@ -57,15 +60,12 @@
     }
   </script>
 
+<RemoveModal bind:open={removing} bind:item bind:purge={purging} reason='' />
+<BanInstanceModal bind:open={banningInstance} bind:banned={item.creator.banned} bind:user={item.creator} />
+<BanCommunityModal bind:open={banningCommunity} bind:banned={item.creator_banned_from_community} bind:user={item.creator} bind:community={item.community}/>
+
 <Menu alignment="{$userSettings.uiState.reverseActionBar ? 'top-left' :  'top-right'}" >
-    <Button
-        on:click={toggleOpen}
-        slot="button"
-        color="tertiary-border"
-        size="square-sm"
-        let:toggleOpen
-        {...$$restProps}
-    >
+    <Button on:click={toggleOpen} slot="button" color="tertiary-border" size="square-sm" let:toggleOpen {...$$restProps}>
         <ShieldIcon filled width={14} />
     </Button>
 
@@ -95,7 +95,11 @@
         </MenuButton>
 
         <!---Remove/Restore Comment--->
-        <MenuButton color={item.comment.removed ? 'success' : 'dangerSecondary'} on:click={() => remove(item)}>
+        <MenuButton color={item.comment.removed ? 'success' : 'dangerSecondary'} on:click={() => {
+                purging = false
+                removing=true
+            }}
+        >
             <Icon src={Trash} size="16" mini />
                 {item.comment.removed ? 'Restore Comment' : 'Remove Comment'}
         </MenuButton>
@@ -104,9 +108,7 @@
         {#if $profile?.user && $profile.user?.local_user_view.person.id != item.creator.id}
             <MenuButton
                 color={item.creator_banned_from_community ? 'success' : 'dangerSecondary'}
-                on:click={() =>
-                    ban(item.creator_banned_from_community, item.creator, item.community)
-                }
+                on:click={() => { banningCommunity = true }}
             >
                 <Icon src={ShieldExclamation} size="16" mini />
                 {item.creator_banned_from_community
@@ -118,7 +120,11 @@
     {/if}
 
     {#if $profile?.user && isAdmin($profile.user)}
-        <MenuButton color="dangerSecondary" on:click={() => remove(item, true)}>
+        <MenuButton color="dangerSecondary" on:click={() => {
+                purging = true
+                removing = true
+            }}
+        >
             <Icon src={Fire} size="16" mini slot="icon" />
             Purge Comment
         </MenuButton>
@@ -127,9 +133,7 @@
         {#if item.creator.id != $profile.user.local_user_view.person.id}
             <MenuButton
                 color={item.creator.banned ? 'success' : 'dangerSecondary'}
-                on:click={() =>
-                    ban(item.creator.banned, item.creator)
-                }
+                on:click={() => { banningInstance = true }}
             >
                 <Icon slot="icon" mini size="16" src={ShieldExclamation} />
                 {item.creator.banned ? 'Unban from Instance' : 'Ban from Instance'}